# vip-portal/src/main/webapp/api-doc/internal-api-doc.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost/internal",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/resources/{id}" : {
      "get" : {
        "tags" : [ "resource-controller" ],
        "operationId" : "get",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "resource-controller" ],
        "operationId" : "createOrUpdate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Resource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "resource-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/applications/{id}" : {
      "get" : {
        "tags" : [ "application-controller" ],
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Application"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "application-controller" ],
        "operationId" : "createOrUpdate_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Application"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Application"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "application-controller" ],
        "operationId" : "deleteApplication",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/session" : {
      "get" : {
        "tags" : [ "session-controller" ],
        "operationId" : "getSession",
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Session"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "session-controller" ],
        "operationId" : "createSession",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AuthenticationCredentials"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Session"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "session-controller" ],
        "operationId" : "deleteSession",
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/resources" : {
      "get" : {
        "tags" : [ "resource-controller" ],
        "operationId" : "list",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "quantity",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10,
            "exclusiveMinimum" : 0,
            "maximum" : 50
          }
        }, {
          "name" : "group",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PrecisePageResource"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "resource-controller" ],
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Resource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      }
    },
    "/applications" : {
      "get" : {
        "tags" : [ "application-controller" ],
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "quantity",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10,
            "exclusiveMinimum" : 0,
            "maximum" : 50
          }
        }, {
          "name" : "group",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PrecisePageApplication"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "application-controller" ],
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Application"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Application"
                }
              }
            }
          }
        }
      }
    },
    "/tags" : {
      "get" : {
        "tags" : [ "tag-controller" ],
        "operationId" : "list_2",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "quantity",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10,
            "exclusiveMinimum" : 0,
            "maximum" : 50
          }
        } ],
        "responses" : {
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorCodeAndMessage"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PrecisePageTag"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "ErrorCodeAndMessage" : {
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "errorMessage" : {
            "type" : "string"
          }
        },
        "required" : [ "errorCode", "errorMessage" ]
      },
      "Group" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "publicGroup" : {
            "type" : "boolean"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "APPLICATION", "RESOURCE" ]
          },
          "auto" : {
            "type" : "boolean"
          }
        }
      },
      "Resource" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "status" : {
            "type" : "boolean"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "LOCAL", "BATCH", "KUBERNETES", "DIRAC" ]
          },
          "configuration" : {
            "type" : "string"
          },
          "engines" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "groups" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Group"
            },
            "uniqueItems" : true
          },
          "groupsNames" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          }
        }
      },
      "Application" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "citation" : {
            "type" : "string"
          },
          "owner" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "groups" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Group"
            },
            "uniqueItems" : true
          },
          "note" : {
            "type" : "string"
          },
          "groupsNames" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          }
        }
      },
      "AuthenticationCredentials" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        },
        "required" : [ "password", "username" ]
      },
      "Session" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          },
          "userlevel" : {
            "type" : "string",
            "enum" : [ "Beginner", "Advanced", "Developer", "Administrator" ]
          }
        }
      },
      "PrecisePageTag" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Tag"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Tag" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "STRING", "BOOLEAN" ]
          },
          "application" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          },
          "visible" : {
            "type" : "boolean"
          },
          "boutiques" : {
            "type" : "boolean"
          }
        }
      },
      "PrecisePageResource" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Resource"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "PrecisePageApplication" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Application"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    }
  }
}